move ThreadView to Linux specific file
- This step is part of the move of linux specific functionality to
separate file in preparation for decoupling the linux Process specific
logic from the output of ELF dump file
