+++ elfcore/src/linux/mod.rs
@@ -1,0 +1,4 @@
+//! This module contains the functionality for
+//! gathering process information on Linux systems.
+
+pub mod process;
